Refactor views into a package
This better matches the other applications and makes it more obvious
what a modules contain views.
