+++ app/main/__init__.py
@@ -3,7 +3,8 @@ from flask import Blueprint
 
 main = Blueprint('main', __name__)
 
-from . import views, service_update_audits, errors
+from . import errors
+from .views import login, service_updates, services
 
 
 @main.after_request
